requirements.txt
Django===1.10.2
psycopg2===2.6.2
pytz===2016.7
channels===0.17.2
PyJWT===1.4.2
